Fix tests and dedupe publish commands
